requirements.txt
pyttsx3
SpeechRecognition
wikipedia
pyautogui
python-dotenv
openai
pyaudio